Adjust feeds for config and about
Removed dependency on about and config in feeds. The services are in core and will get built.
Merged config package into alljoyn
To keep it consistant with about, config package is now part of alljoyn feed.
Fix install paths for config package
Added config as a dep to alljoyn samples
Error in about package name

Change-Id: I3d917c55ef5f5adb2ac73d168e3b33e9ae545b0c
Signed-off-by: Jarrod Moldrich <jarrod@two-bulls.com>